o allow acceleration and steps/mm to be float values.

+++ gcode-machine-control.h
@@ -19,7 +19,7 @@
 
 /* Configuration constants for the controller. */
 struct MachineControlConfig {
-  int axis_steps_per_mm[8];    // Steps per mm for each axis.
+  float axis_steps_per_mm[8];  // Steps per mm for each axis.
   float acceleration;          // Euclidian (x/y) acceleration in mm/s^2
   float max_feedrate;          // Maximum allowed euclidean feedrate in mm/s
   float speed_factor;          // Should be 1.0 by default.
